Added initial menu to Shop Scene Renamed ShopScene to ShopMainScene Stubbed out ShoptSellToPlayerScene
